Addition of the Olimpia Splendid Maestro
Reverse engineered - can evaluate some of the captures here:
https://gist.github.com/yincrash/4e687cbfda9d91f157eefe3b21966113

Product page:
https://olimpiasplendidusa.com/maestro-pro-inverter-12-hp

Changed the default ESP8266 pin to 2 (D4 on nodemcu-styleboards)
Pin 5 is used for the i2c bus and can be occupied on boards with screens
builtin.

Added an additional comment to describe that by default we're treating
the bitstream as LSB to MSB byte ordering.